requirements.txt
# Gmail API and authentication
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# Data processing and parsing
PyYAML>=6.0
beautifulsoup4>=4.9.0
python-dateutil>=2.8.0

# AI service integrations
openai>=1.0.0
anthropic>=0.7.0

# HTTP requests
requests>=2.28.0

# Environment variable loading (optional)
python-dotenv>=0.19.0